+++ lib/graphql/execution/interpreter/runtime/graphql_result.rb
@@ -30,6 +30,8 @@ module GraphQL
           # TODO test full path in Partial
           attr_writer :base_path
 
+          attr_reader :graphql_dead
+
           def current_result
             self
           end
@@ -53,7 +55,14 @@ module GraphQL
             end
           end
 
-          attr_accessor :graphql_dead, :was_scoped
+          def depth
+            @depth ||= if @graphql_parent
+              @graphql_parent.depth + 1
+            else
+              1
+            end
+          end
+
           attr_reader :graphql_parent, :graphql_result_name, :graphql_is_non_null_in_parent,
             :graphql_application_value, :graphql_result_type, :graphql_selections, :graphql_is_eager, :ast_node, :graphql_arguments, :graphql_field