+++ spec/graphql/dataloader_spec.rb
@@ -940,6 +940,7 @@ describe GraphQL::Dataloader do
           query_str = "{ cookbooks { featuredRecipe { name } } }"
           context = { batched_calls_counter: BatchedCallsCounter.new }
           result = schema.execute(query_str, context: context)
+          assert_equal ["Cornbread", "Grits"], result["data"]["cookbooks"].map { |c| c["featuredRecipe"]["name"] }
           refute result.key?("errors")
           assert_equal 1, context[:batched_calls_counter].count
         end
@@ -1190,17 +1191,17 @@ describe GraphQL::Dataloader do
 
             res = schema.execute(query_str, context: { dataloader: fiber_counting_dataloader_class.new })
             assert_nil res.context.dataloader.fiber_limit
-            assert_equal 12, FiberCounting.last_spawn_fiber_count
+            assert_equal 10, FiberCounting.last_spawn_fiber_count
             assert_last_max_fiber_count(9, "No limit works as expected")
 
             res = schema.execute(query_str, context: { dataloader: fiber_counting_dataloader_class.new(fiber_limit: 4) })
             assert_equal 4, res.context.dataloader.fiber_limit
-            assert_equal 14, FiberCounting.last_spawn_fiber_count
+            assert_equal 12, FiberCounting.last_spawn_fiber_count
             assert_last_max_fiber_count(4, "Limit of 4 works as expected")
 
             res = schema.execute(query_str, context: { dataloader: fiber_counting_dataloader_class.new(fiber_limit: 6) })
             assert_equal 6, res.context.dataloader.fiber_limit
-            assert_equal 10, FiberCounting.last_spawn_fiber_count
+            assert_equal 8, FiberCounting.last_spawn_fiber_count
             assert_last_max_fiber_count(6, "Limit of 6 works as expected")
           end