+++ spec/graphql/types/iso_8601_date_spec.rb
@@ -157,7 +157,7 @@ describe GraphQL::Types::ISO8601Date do
       assert_equal expected_errors, parse_date("xyz").map { |e| e["message"] }
       assert_equal expected_errors, parse_date(nil).map { |e| e["message"] }
       assert_equal expected_errors, parse_date([1, 2, 3]).map { |e| e["message"] }
-      assert_equal "A type error was raised: Date cannot be parsed: blah. \nDate must be be able to be parsed as a Ruby Date object.", parse_date("blah", context: { raise_type_error: true })[0]["extensions"]["problems"][0]["explanation"].strip
+      assert_equal "A type error was raised: Date cannot be parsed: blah. \nDate must be able to be parsed as a Ruby Date object.", parse_date("blah", context: { raise_type_error: true })[0]["extensions"]["problems"][0]["explanation"].strip
       assert_equal "Could not coerce value \"blah\" to ISO8601Date", parse_date("blah", context: { raise_type_error: false })[0]["extensions"]["problems"][0]["explanation"]
     end